--- src/pages/DeliveryPage/DeliveryPage.module.scss
@@ -1,15 +1,0 @@
-@use '../../shared/styles/utils/variables' as var;
-
-.help {
-  align-self: flex-start;
-
-  &__wrapper {
-    display: flex;
-    max-width: 100%;
-    gap: 30px;
-  }
-}
-
-.delivery {
-  width: 800px;
-}
