+++ src/pages/HelpPage/HelpPage.tsx
@@ -1,9 +1,13 @@
-import Heading, { HeadingType } from '@/shared/ui/Heading/Heading'
 import styles from './HelpPage.module.scss'
+import Heading, { HeadingType } from '@/shared/ui/Heading/Heading'
 import Subheading from '@/shared/ui/Subheading/Subheading'
 import HelpCategories from '@/widgets/HelpCategories/HelpCategories'
 import FormReturn from '@/widgets/FormReturn/FormReturn'
 import WrapperForMainContent from '../../components/WrapperForMainContent/WrapperForMainContent'
+
+// @TODO: Сверстать форму возврата
+// Сделал страницу помощи
+// https://github.com/Studio-Yandex-Practicum/maxboom_frontend/issues/222
 
 const HelpPage = () => {
   return (
