+++ src/shared/model/types/common.ts
@@ -27,3 +27,17 @@ export type TOrder = {
   productsSumWithoutDiscount: number
   total: number
 }
+
+export type TFormReturn = {
+  name: string
+  surname: string
+  email: string
+  tel: string
+  numberOrder: string
+  dateOrder: string
+  itemInfo: string
+  model: string
+  amount: string
+  textArea: string
+  myChoice: string
+}
