+++ src/widgets/FormReturn/FormReturn.tsx
@@ -1,31 +1,18 @@
-import styles from './FormReturn.module.scss'
-import Heading, { HeadingType } from '@/shared/ui/Heading/Heading'
 import { Field, ErrorMessage, Formik, Form } from 'formik'
-import { Input } from '@/shared/ui/Input/Input'
-import { validationSchema } from '../../features/login/model/validation/validation'
 import classNames from 'classnames'
+import styles from './FormReturn.module.scss'
+import { validationSchema } from '@/features/login/model/validation/validation'
+import Heading, { HeadingType } from '@/shared/ui/Heading/Heading'
+import { Input } from '@/shared/ui/Input/Input'
 import { Textarea } from '@/shared/ui/Textarea/Textarea'
-import { Button, ButtonTheme, ButtonDesign, ButtonSize } from '../../shared/ui/Button/Button'
-import Label from '../../shared/ui/Label/Label'
-import Checkbox from '../../shared/ui/Checkbox/Checkbox'
+import { Button, ButtonTheme, ButtonDesign, ButtonSize } from '@/shared/ui/Button/Button'
+import Label from '@/shared/ui/Label/Label'
+import Checkbox from '@/shared/ui/Checkbox/Checkbox'
+import { TFormReturn } from '@/shared/model/types/common'
 
 // @TODO: Сверстать форму возврата
 // Сделал checkbox
 // https://github.com/Studio-Yandex-Practicum/maxboom_frontend/issues/222
-
-export type TFormReturn = {
-  name: string
-  surname: string
-  email: string
-  tel: string
-  numberOrder: string
-  dateOrder: string
-  itemInfo: string
-  model: string
-  amount: string
-  textArea: string
-  myChoice: string
-}
 
 const initialValues: TFormReturn = {
   name: '',
